added logMetrics fns

+++ headers/QueuePair.h
@@ -10,8 +10,8 @@ struct QueuePair {
 
     ReadyQueue* rq;
     Processor* prc;
-    int active_count;
-    int sample_queue;
+    float active_count;
+    float sample_queue;
 };
